feat(validations): alteracoes para garantir validacoes - altera tipo de id de Long para UUID em Role - cria exception para senha invalida - remove mock e impl logica para evitar remoção do ADMIN - adiciona dados do ADMIN em variaveis de ambiente

+++ src/main/java/br/com/barbosa/services/UserService.java
@@ -3,6 +3,7 @@ package br.com.barbosa.services;
 import br.com.barbosa.entities.Role;
 import br.com.barbosa.entities.User;
 import br.com.barbosa.exceptions.EmailAlreadyExistsException;
+import br.com.barbosa.exceptions.PasswordValidationException;
 import br.com.barbosa.exceptions.ResourceNotFoundException;
 import br.com.barbosa.repositories.RoleRepository;
 import br.com.barbosa.repositories.UserRepository;
@@ -11,6 +12,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.UUID;
 
 @Service
 public class UserService {
@@ -24,7 +26,7 @@ public class UserService {
     @Autowired
     private BCryptPasswordEncoder passwordEncoder;
 
-    public User findById(Long id) {
+    public User findById(UUID id) {
         return repository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Usuário com ID " + id + " não encontrado."));
     }
@@ -38,6 +40,8 @@ public class UserService {
             throw new EmailAlreadyExistsException("O e-mail informado já está cadastrado. Por favor, utilize outro e-mail.");
         }
 
+        validatePasswordLength(user.getPassword());
+
         user.setPassword(passwordEncoder.encode(user.getPassword()));
 
         Role userRole = roleRepository.findByRoleName("ROLE_USER")
@@ -47,9 +51,23 @@ public class UserService {
         return repository.save(user);
     }
 
-    public void deleteById(Long id) {
+    private void validatePasswordLength(String password) {
+        if (password.length() < 8 || password.length() > 50) {
+            throw new PasswordValidationException("A senha deve ter entre 8 e 50 caracteres.");
+        }
+    }
+
+    public void deleteById(UUID id) {
         User user = repository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Usuário com ID " + id + " não encontrado."));
+
+        boolean isAdmin = user.getRoles().stream()
+                .anyMatch(role -> "ROLE_ADMIN".equals(role.getRoleName()));
+
+        if (isAdmin) {
+            throw new RuntimeException("Não é possível excluir o usuário administrador.");
+        }
+
         repository.delete(user);
     }
 }
